Adjust error reporting, tweak `or/p`, and add `lookahead/p`
This is a fairly substantial change to parts of the megaparsack
internals, all of which was triggered by my attempt to just add
`lookahead/p`. Prior to this commit, the implementation used a rather
strange definition of `or/p` from the original paper, which continued to
try parsers after a successful parse that consumed no input.

This behavior is confusing, wasteful of time and space, and makes it
somewhat tricky to pin down the precise semantics of which errors are
reported when. What’s more, no existing Parsec implementation I can find
actually implements it---even the parsec library itself switched to the
“first success” strategy over a decade ago. Technically this is a
breaking change, but I have a hunch that any programs actually affected
probably still prefer the new behavior to the old one.

In addition to cleaning up `or/p`, this commit drastically tidies up the
logic around selection and merging of parse errors. Errors are now only
merged if they originate from the same location, and that location is
determined based on an index into the token stream, so it is reliable
even in the face of nonsense source locations. Errors are also merged
more consistently, always preferring errors that made more progress,
rather than sometimes preferring errors that appear later and sometimes
dropping errors entirely.

Finally, this commit also adds `lookahead/p`, which behaves the way one
would expect based on existing Parsec implementations and is by far the
most straightforward part of this change.