Add `noncommittal/p` to allow backtracking after a successful parse
This is inspired by a discussion on the Racket Discord server that
involved a use of `many/p` with a #:sep argument along these lines:

    (many/p p #:sep (many+/p spaces/p))

The user noted that trailing spaces after the sequence will result in a
parse error, since the separator parser will consume them, and that will
cause `many/p` to commit to parsing an additional element.

Such a problem can be fixed by adjusting the separator parser so that it
is not considered to have consumed input upon success, which is
effectively the dual behavior to `try/p`. No parsec-style library
provides such a combinator to my knowledge, but I don’t see anything
wrong with it, so I have added it as `noncommittal/p`. This can be used
to fix the original example by writing

    (many/p p #:sep (noncommittal/p (many+/p spaces/p)))

which allows a failure in `p` to backtrack past a sequence of spaces
consumed by the separator parser.